Use rootless docker in docker to construct RDS
TSan is used as an prototype.

First of all, the official rootless dind image is running as a service provider to offer the rootless docker daemon to RDS. This container itself runs in the regular root mode but it's local in the host. The public users have no access to this container.

The new architecture of RDS has two layers of docker.

The outer docker is built upon the official docker image, which is based on Alpine Linux and supports rootless docker very well. The Flask server and its dependencies are installed here. It receives API calls and create a sandbox,the inner docker.

The inner docker is a highly customized docker image according to our needs. It runs in rootless mode without network ability. The input file is copied into the sandbox and the tool will compile the code and detect the data race. Then the outer docker will copy the output out of inner docker and destroys the sandbox.

All the executables are isolated in the sandbox. Even the sandbox is hacked, it won't be harmful because it's running as non-root user and has no access to any network. The output from the sandbox is considered as plain text file, which will be never executed in the outer docker.