Pull the tool docker image while starting the TSan server
It will get the tool ready at the very beginning.
Otherwise, the user who submits the first task has to wait for a very long time to wait until the tool docker image is pulled.
The followed users won't face this issue because by then the tool docker image is always there.